rename server -> service
