+++ ee/tabby-webserver/src/service/auth.rs
