+++ ee/tabby-webserver/src/service/worker.rs
